+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0 / 2013-08-22
+
+Production-ready release
+
+-  Only set years-old cache-friendly last-modified header on assets with max-age of at least 1 month
+
 ## 0.1.0 / 2013-08-21
 
 Initial release